requirements.txt
click==6.7
eventlet==0.23.0
Flask==1.0.2
Flask-MQTT==1.0.4
Flask-SocketIO==3.0.1
greenlet==0.4.14
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
paho-mqtt==1.3.1
python-engineio==2.2.0
python-socketio==2.0.0
six==1.11.0
typing==3.6.4
Werkzeug==0.14.1
